+++ stroom-core/src/main/java/stroom/core/query/SuggestionsService.java
@@ -5,6 +5,5 @@ import stroom.query.shared.FetchSuggestionsRequest;
 import java.util.List;
 
 public interface SuggestionsService {
-
     List<String> fetch(final FetchSuggestionsRequest request);
 }