+++ stroom-proxy/stroom-proxy-remote-api/src/main/java/stroom/proxy/StroomStatusCode.java
@@ -17,6 +17,9 @@ public enum StroomStatusCode {
     FEED_IS_NOT_SET_TO_RECEIVED_DATA(HttpServletResponse.SC_NOT_ACCEPTABLE, 110, "Feed is not set to receive data",
             "The feed you have provided has not been setup to receive data"),
 
+    UNEXPECTED_DATA_TYPE(HttpServletResponse.SC_NOT_ACCEPTABLE, 120, "Unexpected data type",
+            "The data type supplied is not expected"),
+
     UNKNOWN_COMPRESSION(HttpServletResponse.SC_NOT_ACCEPTABLE, 200, "Unknown compression",
             "Compression argument must be one of ZIP, GZIP and NONE"),