+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#1665** : Recognised meta types can now be specified in config and drop downs now allow selection in the pipeline editor.
+
 * Issue **#1793** : Fixed Solr search query creation.
 
 * Issue **#1791** : Fixed Solr connection test response.